requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pymongo>=4.0.0
dvc>=3.0.0
prometheus_client>=0.16.0
pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0.0
scikit-learn>=1.0.0
python-gitlab>=3.0.0
python-dotenv>=0.19.0
mlflow>=2.8.0
jinja2>=3.0.0